pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "semantic-reducer"
version = "0.1.0"
description = "O(1) semantic vocabulary reduction using BERT and FAISS."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch>=1.10.0",
    "numpy>=1.21.0",
    "faiss-cpu>=1.7.0",
    "tqdm>=4.60.0",
    "transformers>=4.20.0"
]